test(e2e): fold the early-bound test into pluginAcceptance and retire pluginLifecycle (#143)
The two plugin service-principal suites ran the same flow twice — one through the
command palette, one through the panel buttons — for ~15 minutes of duplicate wall
clock per run. Only one test was unique to pluginLifecycle: early-bound generation,
which matters because the build immediately after it is what proves the generated
classes still compile into the package (#130).

So the early-bound test moves into pluginAcceptance, driven by its own panel button
like everything else in that suite, seeded with a modelbuilder.json so it runs pac
directly instead of opening the config wizard. It keeps the assertion that made it
worth having: pac exits 0 on failure, so the log line is only the command's claim —
generated/*.cs on disk is the fact.

pluginLifecycle is deleted; TESTING.md's suite inventory and CLAUDE.md's suite list
follow. (test/live/pluginLifecycle.spec.ts is a different, still-live file.)
